Fix holonomic.cpp so that it doens't output all zeros :0

+++ blueshift_core/src/holonomic/holonomic.cpp
@@ -7,14 +7,14 @@ blueshift_interfaces::msg::Motors holonomic_math(double lx, double ly, double lz
 {
     auto m = blueshift_interfaces::msg::Motors();
 
-    double bottom_scale = 1;
-    double top_scale = 1;
+    double bottom_scale = (double) 1;
+    double top_scale = (double) 1;
 
     // limiter_type = 0 for clamp limiter, 1 for 1/3 limiter and 2 for dynamic limiter
     if (limiter_type == 1)
     {
-        bottom_scale = 1 / 3;
-        top_scale = 1 / 3;
+        bottom_scale = (double) 1 / 3;
+        top_scale = (double) 1 / 3;
     }
     if (limiter_type == 2)
     {
